% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGElement_methods.R
\name{add_method}
\alias{add_method}
\title{Add a method to the SVGElement class which creates a new child SVGElement of the specified type}
\usage{
add_method(method_name, element_name, args)
}
\arguments{
\item{method_name}{name of method in SVGElement class}

\item{element_name}{the name of the SVG element to add}

\item{args}{character vector of argument names}
}
\description{
Add a method to the SVGElement class which creates a new child SVGElement of the specified type
}
