% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGDocument.R
\name{as.character.SVGDocument}
\alias{as.character.SVGDocument}
\title{Character representation of SVGDocument}
\usage{
\method{as.character}{SVGDocument}(x, include_declaration = TRUE, ...)
}
\arguments{
\item{x}{SVGDocument}

\item{include_declaration}{Include the SVG declaration at the top? Default: TRUE}

\item{...}{other arguments}
}
\description{
Character representation of SVGDocument
}
