% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGElement.R
\name{as.character.SVGElement}
\alias{as.character.SVGElement}
\title{Character representation of SVGElement}
\usage{
\method{as.character}{SVGElement}(x, ...)
}
\arguments{
\item{x}{SVGElement}

\item{...}{other arguments}
}
\description{
Character representation of SVGElement
}
