% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGLiteral.R
\name{as.character.SVGLiteral}
\alias{as.character.SVGLiteral}
\title{Character representation of SVGLiteral}
\usage{
\method{as.character}{SVGLiteral}(x, ...)
}
\arguments{
\item{x}{SVGLiteral}

\item{...}{other arguments}
}
\description{
Character representation of SVGLiteral
}
