% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGNode.R
\name{as.character.SVGNode}
\alias{as.character.SVGNode}
\title{Character representation of SVGNode}
\usage{
\method{as.character}{SVGNode}(x, ...)
}
\arguments{
\item{x}{SVGNode}

\item{...}{other arguments}
}
\description{
Character representation of SVGNode
}
