% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGDocument.R
\name{knit_print.SVGDocument}
\alias{knit_print.SVGDocument}
\title{knitr/rmarkdown compatibility}
\usage{
knit_print.SVGDocument(x, ...)
}
\arguments{
\item{x}{SVGDocument}

\item{...}{other arguments}
}
\description{
knitr/rmarkdown compatibility
}
