% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGFilter.R
\name{knit_print.SVGFilter}
\alias{knit_print.SVGFilter}
\title{knitr/rmarkdown compatibility}
\usage{
knit_print.SVGFilter(x, ...)
}
\arguments{
\item{x}{SVGFilter}

\item{...}{other arguments}
}
\description{
knitr/rmarkdown compatibility
}
