% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGPattern.R
\name{knit_print.SVGPattern}
\alias{knit_print.SVGPattern}
\title{knitr/rmarkdown compatibility}
\usage{
knit_print.SVGPattern(x, ...)
}
\arguments{
\item{x}{SVGPattern}

\item{...}{other arguments}
}
\description{
knitr/rmarkdown compatibility
}
