% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGDocument.R
\name{show_svg}
\alias{show_svg}
\title{SVG shower}
\usage{
show_svg(svg, viewer = getOption("viewer", utils::browseURL))
}
\arguments{
\item{svg}{SVG text or object}

\item{viewer}{viewer}
}
\description{
SVG shower
}
