% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stag.R
\name{stag}
\alias{stag}
\title{SVG helper}
\description{
SVG builder functions.  Similar in purpose to \code{shiny::tags},
but with auto-complete of attribute names as part of the function call.
}
\examples{
\dontrun{
stag$circle(cx = 10, cy = 10, r = 15)
# <circle cx="10" cy="10" r="15" />
}


}
