% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_prop.R
\docType{data}
\name{svg_prop}
\alias{svg_prop}
\title{SVG property helper}
\format{
An object of class \code{list} of length 42.
}
\usage{
svg_prop
}
\description{
Uses autocomplete to help write some standard propertys
}
\keyword{datasets}
