% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg2png.R
\name{svg_to_png}
\alias{svg_to_png}
\title{Render SVG to PNG}
\usage{
svg_to_png(svg, opts = list(), output = "output.png")
}
\arguments{
\item{svg}{string or raw vector with svg image}

\item{opts}{list with rendering options}

\item{output}{filename of png file}
}
\description{
Converts an svg string to a PNG image using resvg.
}
\examples{
svg <- readLines("https://upload.wikimedia.org/wikipedia/commons/f/fd/Ghostscript_Tiger.svg")
svg_to_png(svg, output = 'tiger.png')
# browseURL('tiger.png')
unlink('tiger.png')
}
