% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGPattern.R
\name{SVGPattern}
\alias{SVGPattern}
\alias{svg_pattern}
\title{SVGPattern Class}
\description{
SVGPattern Class

SVGPattern Class
}
\details{
This is a slightly specialized subclass of \code{SVGElement} which has methods
to specifically handle SVG \code{<pattern>} nodes
}
\section{Super class}{
\code{\link[minisvg:SVGElement]{minisvg::SVGElement}} -> \code{SVGPattern}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SVGPattern$new()}}
\item \href{#method-as_full_svg}{\code{SVGPattern$as_full_svg()}}
\item \href{#method-save_full_svg}{\code{SVGPattern$save_full_svg()}}
\item \href{#method-show}{\code{SVGPattern$show()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="a">}\href{../../minisvg/html/SVGElement.html#method-a}{\code{minisvg::SVGElement$a()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="add">}\href{../../minisvg/html/SVGElement.html#method-add}{\code{minisvg::SVGElement$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="animate">}\href{../../minisvg/html/SVGElement.html#method-animate}{\code{minisvg::SVGElement$animate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="animateMotion">}\href{../../minisvg/html/SVGElement.html#method-animateMotion}{\code{minisvg::SVGElement$animateMotion()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="animateTransform">}\href{../../minisvg/html/SVGElement.html#method-animateTransform}{\code{minisvg::SVGElement$animateTransform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="append">}\href{../../minisvg/html/SVGElement.html#method-append}{\code{minisvg::SVGElement$append()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="as_character">}\href{../../minisvg/html/SVGElement.html#method-as_character}{\code{minisvg::SVGElement$as_character()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="as_character_inner">}\href{../../minisvg/html/SVGElement.html#method-as_character_inner}{\code{minisvg::SVGElement$as_character_inner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="circle">}\href{../../minisvg/html/SVGElement.html#method-circle}{\code{minisvg::SVGElement$circle()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="clipPath">}\href{../../minisvg/html/SVGElement.html#method-clipPath}{\code{minisvg::SVGElement$clipPath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="copy">}\href{../../minisvg/html/SVGElement.html#method-copy}{\code{minisvg::SVGElement$copy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="defs">}\href{../../minisvg/html/SVGElement.html#method-defs}{\code{minisvg::SVGElement$defs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="desc">}\href{../../minisvg/html/SVGElement.html#method-desc}{\code{minisvg::SVGElement$desc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="discard">}\href{../../minisvg/html/SVGElement.html#method-discard}{\code{minisvg::SVGElement$discard()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="ellipse">}\href{../../minisvg/html/SVGElement.html#method-ellipse}{\code{minisvg::SVGElement$ellipse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="filter">}\href{../../minisvg/html/SVGElement.html#method-filter}{\code{minisvg::SVGElement$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="foreignObject">}\href{../../minisvg/html/SVGElement.html#method-foreignObject}{\code{minisvg::SVGElement$foreignObject()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="g">}\href{../../minisvg/html/SVGElement.html#method-g}{\code{minisvg::SVGElement$g()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="image">}\href{../../minisvg/html/SVGElement.html#method-image}{\code{minisvg::SVGElement$image()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="line">}\href{../../minisvg/html/SVGElement.html#method-line}{\code{minisvg::SVGElement$line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="linearGradient">}\href{../../minisvg/html/SVGElement.html#method-linearGradient}{\code{minisvg::SVGElement$linearGradient()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="marker">}\href{../../minisvg/html/SVGElement.html#method-marker}{\code{minisvg::SVGElement$marker()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="mask">}\href{../../minisvg/html/SVGElement.html#method-mask}{\code{minisvg::SVGElement$mask()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="matrix">}\href{../../minisvg/html/SVGElement.html#method-matrix}{\code{minisvg::SVGElement$matrix()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="mpath">}\href{../../minisvg/html/SVGElement.html#method-mpath}{\code{minisvg::SVGElement$mpath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="path">}\href{../../minisvg/html/SVGElement.html#method-path}{\code{minisvg::SVGElement$path()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="pattern">}\href{../../minisvg/html/SVGElement.html#method-pattern}{\code{minisvg::SVGElement$pattern()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="polygon">}\href{../../minisvg/html/SVGElement.html#method-polygon}{\code{minisvg::SVGElement$polygon()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="polyline">}\href{../../minisvg/html/SVGElement.html#method-polyline}{\code{minisvg::SVGElement$polyline()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="print">}\href{../../minisvg/html/SVGElement.html#method-print}{\code{minisvg::SVGElement$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="radialGradient">}\href{../../minisvg/html/SVGElement.html#method-radialGradient}{\code{minisvg::SVGElement$radialGradient()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="rect">}\href{../../minisvg/html/SVGElement.html#method-rect}{\code{minisvg::SVGElement$rect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="remove">}\href{../../minisvg/html/SVGElement.html#method-remove}{\code{minisvg::SVGElement$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="reset_transform">}\href{../../minisvg/html/SVGElement.html#method-reset_transform}{\code{minisvg::SVGElement$reset_transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="rotate">}\href{../../minisvg/html/SVGElement.html#method-rotate}{\code{minisvg::SVGElement$rotate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="save">}\href{../../minisvg/html/SVGElement.html#method-save}{\code{minisvg::SVGElement$save()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="scale">}\href{../../minisvg/html/SVGElement.html#method-scale}{\code{minisvg::SVGElement$scale()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="script">}\href{../../minisvg/html/SVGElement.html#method-script}{\code{minisvg::SVGElement$script()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="set">}\href{../../minisvg/html/SVGElement.html#method-set}{\code{minisvg::SVGElement$set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="skewX">}\href{../../minisvg/html/SVGElement.html#method-skewX}{\code{minisvg::SVGElement$skewX()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="skewY">}\href{../../minisvg/html/SVGElement.html#method-skewY}{\code{minisvg::SVGElement$skewY()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="stop">}\href{../../minisvg/html/SVGElement.html#method-stop}{\code{minisvg::SVGElement$stop()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="style">}\href{../../minisvg/html/SVGElement.html#method-style}{\code{minisvg::SVGElement$style()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="switch">}\href{../../minisvg/html/SVGElement.html#method-switch}{\code{minisvg::SVGElement$switch()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="symbol">}\href{../../minisvg/html/SVGElement.html#method-symbol}{\code{minisvg::SVGElement$symbol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="text">}\href{../../minisvg/html/SVGElement.html#method-text}{\code{minisvg::SVGElement$text()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="textPath">}\href{../../minisvg/html/SVGElement.html#method-textPath}{\code{minisvg::SVGElement$textPath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="title">}\href{../../minisvg/html/SVGElement.html#method-title}{\code{minisvg::SVGElement$title()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="translate">}\href{../../minisvg/html/SVGElement.html#method-translate}{\code{minisvg::SVGElement$translate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="update">}\href{../../minisvg/html/SVGElement.html#method-update}{\code{minisvg::SVGElement$update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="minisvg" data-topic="SVGElement" data-id="use">}\href{../../minisvg/html/SVGElement.html#method-use}{\code{minisvg::SVGElement$use()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialise an SVGPattern object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGPattern$new(..., name = "pattern")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed to \code{SVGElement$new()}}

\item{\code{name}}{defaults to 'pattern', but some gradients may also be used here}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_full_svg"></a>}}
\subsection{Method \code{as_full_svg()}}{
Wrap the SVG for this pattern in a full SVG document and return the text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGPattern$as_full_svg(width = 400, height = width)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height, width}}{dimensions of SVG wrapper around this pattern}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_full_svg"></a>}}
\subsection{Method \code{save_full_svg()}}{
Save the SVG for this pattern in a full SVG document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGPattern$save_full_svg(filename, include_declaration = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename for output}

\item{\code{include_declaration}}{Include leading XML declaration. default: TRUE}

\item{\code{...}}{Further arguments passed to \code{SVGPattern$as_full_svg()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\subsection{Method \code{show()}}{
Render this pattern in the context of a complete SVG document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGPattern$show(..., viewer = getOption("viewer", utils::browseURL))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed to \code{SVGPattern$save_full_svg()}}

\item{\code{viewer}}{viewer.}
}
\if{html}{\out{</div>}}
}
}
}
