% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-patterns.R
\name{SVGPatternList_to_svg}
\alias{SVGPatternList_to_svg}
\title{Create a complete SVG object from an svg pattern}
\usage{
SVGPatternList_to_svg(pattern_list, width = 400, height = 400, ncol = 2, ...)
}
\arguments{
\item{pattern_list}{svg_pattern_list object}

\item{width, height}{the display width of the surrounding SVG wrapper. defualt: 400x400}

\item{ncol}{number of columns. if NULL, then will use an auto-layout}

\item{...}{other arguments ignored}
}
\description{
Create a complete SVG object from an svg pattern
}
