% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGElement.R
\name{SVGElement}
\alias{SVGElement}
\alias{svg_elem}
\title{A class representing a single SVG element.}
\description{
A class representing a single SVG element.

A class representing a single SVG element.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Tag name for this node e.g. "circle"}

\item{\code{attribs}}{named list of attributes for this node}

\item{\code{children}}{list of direct child nodes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-a}{\code{SVGElement$a()}}
\item \href{#method-animate}{\code{SVGElement$animate()}}
\item \href{#method-animateMotion}{\code{SVGElement$animateMotion()}}
\item \href{#method-animateTransform}{\code{SVGElement$animateTransform()}}
\item \href{#method-circle}{\code{SVGElement$circle()}}
\item \href{#method-clipPath}{\code{SVGElement$clipPath()}}
\item \href{#method-defs}{\code{SVGElement$defs()}}
\item \href{#method-desc}{\code{SVGElement$desc()}}
\item \href{#method-discard}{\code{SVGElement$discard()}}
\item \href{#method-ellipse}{\code{SVGElement$ellipse()}}
\item \href{#method-feBlend}{\code{SVGElement$feBlend()}}
\item \href{#method-feColorMatrix}{\code{SVGElement$feColorMatrix()}}
\item \href{#method-feComponentTransfer}{\code{SVGElement$feComponentTransfer()}}
\item \href{#method-feComposite}{\code{SVGElement$feComposite()}}
\item \href{#method-feConvolveMatrix}{\code{SVGElement$feConvolveMatrix()}}
\item \href{#method-feDiffuseLighting}{\code{SVGElement$feDiffuseLighting()}}
\item \href{#method-feDisplacementMap}{\code{SVGElement$feDisplacementMap()}}
\item \href{#method-feFlood}{\code{SVGElement$feFlood()}}
\item \href{#method-feFuncR}{\code{SVGElement$feFuncR()}}
\item \href{#method-feFuncG}{\code{SVGElement$feFuncG()}}
\item \href{#method-feFuncB}{\code{SVGElement$feFuncB()}}
\item \href{#method-feFuncA}{\code{SVGElement$feFuncA()}}
\item \href{#method-feGaussianBlur}{\code{SVGElement$feGaussianBlur()}}
\item \href{#method-feImage}{\code{SVGElement$feImage()}}
\item \href{#method-feMerge}{\code{SVGElement$feMerge()}}
\item \href{#method-feMergeNode}{\code{SVGElement$feMergeNode()}}
\item \href{#method-feMorphology}{\code{SVGElement$feMorphology()}}
\item \href{#method-feOffset}{\code{SVGElement$feOffset()}}
\item \href{#method-feSpecularLighting}{\code{SVGElement$feSpecularLighting()}}
\item \href{#method-feTile}{\code{SVGElement$feTile()}}
\item \href{#method-feTurbulence}{\code{SVGElement$feTurbulence()}}
\item \href{#method-filter}{\code{SVGElement$filter()}}
\item \href{#method-foreignObject}{\code{SVGElement$foreignObject()}}
\item \href{#method-g}{\code{SVGElement$g()}}
\item \href{#method-image}{\code{SVGElement$image()}}
\item \href{#method-line}{\code{SVGElement$line()}}
\item \href{#method-linearGradient}{\code{SVGElement$linearGradient()}}
\item \href{#method-marker}{\code{SVGElement$marker()}}
\item \href{#method-mask}{\code{SVGElement$mask()}}
\item \href{#method-mpath}{\code{SVGElement$mpath()}}
\item \href{#method-path}{\code{SVGElement$path()}}
\item \href{#method-pattern}{\code{SVGElement$pattern()}}
\item \href{#method-radialGradient}{\code{SVGElement$radialGradient()}}
\item \href{#method-rect}{\code{SVGElement$rect()}}
\item \href{#method-script}{\code{SVGElement$script()}}
\item \href{#method-set}{\code{SVGElement$set()}}
\item \href{#method-stop}{\code{SVGElement$stop()}}
\item \href{#method-style}{\code{SVGElement$style()}}
\item \href{#method-switch}{\code{SVGElement$switch()}}
\item \href{#method-symbol}{\code{SVGElement$symbol()}}
\item \href{#method-text}{\code{SVGElement$text()}}
\item \href{#method-textPath}{\code{SVGElement$textPath()}}
\item \href{#method-title}{\code{SVGElement$title()}}
\item \href{#method-use}{\code{SVGElement$use()}}
\item \href{#method-matrix}{\code{SVGElement$matrix()}}
\item \href{#method-translate}{\code{SVGElement$translate()}}
\item \href{#method-scale}{\code{SVGElement$scale()}}
\item \href{#method-rotate}{\code{SVGElement$rotate()}}
\item \href{#method-skewX}{\code{SVGElement$skewX()}}
\item \href{#method-skewY}{\code{SVGElement$skewY()}}
\item \href{#method-polygon}{\code{SVGElement$polygon()}}
\item \href{#method-polyline}{\code{SVGElement$polyline()}}
\item \href{#method-new}{\code{SVGElement$new()}}
\item \href{#method-update}{\code{SVGElement$update()}}
\item \href{#method-append}{\code{SVGElement$append()}}
\item \href{#method-add}{\code{SVGElement$add()}}
\item \href{#method-remove}{\code{SVGElement$remove()}}
\item \href{#method-reset_transform}{\code{SVGElement$reset_transform()}}
\item \href{#method-as_character_inner}{\code{SVGElement$as_character_inner()}}
\item \href{#method-as_character}{\code{SVGElement$as_character()}}
\item \href{#method-print}{\code{SVGElement$print()}}
\item \href{#method-save}{\code{SVGElement$save()}}
\item \href{#method-copy}{\code{SVGElement$copy()}}
\item \href{#method-clone}{\code{SVGElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-a"></a>}}
\subsection{Method \code{a()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$a(..., href = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-animate"></a>}}
\subsection{Method \code{animate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$animate(
  ...,
  attributeName = NULL,
  attributeType = NULL,
  from = NULL,
  to = NULL,
  dur = NULL,
  repeatCount = NULL,
  begin = NULL,
  end = NULL,
  calcMode = NULL,
  values = NULL,
  keyTimes = NULL,
  keySplines = NULL,
  by = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-animateMotion"></a>}}
\subsection{Method \code{animateMotion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$animateMotion(
  ...,
  calcMode = NULL,
  path = NULL,
  keyPoints = NULL,
  rotate = NULL,
  origin = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-animateTransform"></a>}}
\subsection{Method \code{animateTransform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$animateTransform(
  ...,
  attributeName = NULL,
  attributeType = NULL,
  by = NULL,
  from = NULL,
  to = NULL,
  type = NULL,
  dur = NULL,
  repeatCount = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-circle"></a>}}
\subsection{Method \code{circle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$circle(..., cx = NULL, cy = NULL, r = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clipPath"></a>}}
\subsection{Method \code{clipPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$clipPath(..., id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-defs"></a>}}
\subsection{Method \code{defs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$defs(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-desc"></a>}}
\subsection{Method \code{desc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$desc(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discard"></a>}}
\subsection{Method \code{discard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$discard(..., begin = NULL, href = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipse"></a>}}
\subsection{Method \code{ellipse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$ellipse(..., cx = NULL, cy = NULL, rx = NULL, ry = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feBlend"></a>}}
\subsection{Method \code{feBlend()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feBlend(..., in_ = NULL, in2 = NULL, mode = NULL, result = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feColorMatrix"></a>}}
\subsection{Method \code{feColorMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feColorMatrix(
  ...,
  in_ = NULL,
  type = NULL,
  values = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feComponentTransfer"></a>}}
\subsection{Method \code{feComponentTransfer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feComponentTransfer(..., in_ = NULL, result = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feComposite"></a>}}
\subsection{Method \code{feComposite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feComposite(
  ...,
  in_ = NULL,
  in2 = NULL,
  operator = NULL,
  k1 = NULL,
  k2 = NULL,
  k3 = NULL,
  k4 = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feConvolveMatrix"></a>}}
\subsection{Method \code{feConvolveMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feConvolveMatrix(
  ...,
  in_ = NULL,
  order = NULL,
  kernelMatrix = NULL,
  divisor = NULL,
  bias = NULL,
  targetX = NULL,
  targetY = NULL,
  edgeMode = NULL,
  kernelUnitLength = NULL,
  preserveAlpha = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feDiffuseLighting"></a>}}
\subsection{Method \code{feDiffuseLighting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feDiffuseLighting(
  ...,
  in_ = NULL,
  surfaceScale = NULL,
  diffuseConstant = NULL,
  kernelUnitLength = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feDisplacementMap"></a>}}
\subsection{Method \code{feDisplacementMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feDisplacementMap(
  ...,
  in_ = NULL,
  in2 = NULL,
  scale = NULL,
  xChannelSelector = NULL,
  yChannelSelector = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feFlood"></a>}}
\subsection{Method \code{feFlood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feFlood(
  ...,
  flood_color = NULL,
  flood_opacity = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feFuncR"></a>}}
\subsection{Method \code{feFuncR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feFuncR(..., type = NULL, tableValues = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feFuncG"></a>}}
\subsection{Method \code{feFuncG()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feFuncG(..., type = NULL, tableValues = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feFuncB"></a>}}
\subsection{Method \code{feFuncB()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feFuncB(..., type = NULL, tableValues = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feFuncA"></a>}}
\subsection{Method \code{feFuncA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feFuncA(..., type = NULL, tableValues = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feGaussianBlur"></a>}}
\subsection{Method \code{feGaussianBlur()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feGaussianBlur(
  ...,
  in_ = NULL,
  stdDeviation = NULL,
  edgeMode = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feImage"></a>}}
\subsection{Method \code{feImage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feImage(
  ...,
  preserveAspectRatio = NULL,
  xlink_href = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feMerge"></a>}}
\subsection{Method \code{feMerge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feMerge(..., result = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feMergeNode"></a>}}
\subsection{Method \code{feMergeNode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feMergeNode(..., in_ = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feMorphology"></a>}}
\subsection{Method \code{feMorphology()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feMorphology(
  ...,
  in_ = NULL,
  operator = NULL,
  radius = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feOffset"></a>}}
\subsection{Method \code{feOffset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feOffset(..., in_ = NULL, dx = NULL, dy = NULL, result = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feSpecularLighting"></a>}}
\subsection{Method \code{feSpecularLighting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feSpecularLighting(
  ...,
  in_ = NULL,
  surfaceScale = NULL,
  specularConstant = NULL,
  specularExponent = NULL,
  kernelUnitLength = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feTile"></a>}}
\subsection{Method \code{feTile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feTile(..., in_ = NULL, result = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feTurbulence"></a>}}
\subsection{Method \code{feTurbulence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$feTurbulence(
  ...,
  type = NULL,
  baseFrequency = NULL,
  numOctaves = NULL,
  seed = NULL,
  stitchTiles = NULL,
  result = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter"></a>}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$filter(
  ...,
  id = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  filterRes = NULL,
  filterUnits = NULL,
  primitiveUnits = NULL,
  xlink_href = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-foreignObject"></a>}}
\subsection{Method \code{foreignObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$foreignObject(..., x = NULL, y = NULL, width = NULL, height = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-g"></a>}}
\subsection{Method \code{g()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$g(..., id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-image"></a>}}
\subsection{Method \code{image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$image(
  ...,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  xlink_href = NULL,
  preserveAspectRatio = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-line"></a>}}
\subsection{Method \code{line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$line(..., x1 = NULL, y1 = NULL, x2 = NULL, y2 = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linearGradient"></a>}}
\subsection{Method \code{linearGradient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$linearGradient(
  ...,
  x1 = NULL,
  y1 = NULL,
  x2 = NULL,
  y2 = NULL,
  href = NULL,
  gradientTransform = NULL,
  gradientUnits = NULL,
  spreadMethod = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-marker"></a>}}
\subsection{Method \code{marker()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$marker(
  ...,
  refX = NULL,
  refY = NULL,
  markerWidth = NULL,
  markerHeight = NULL,
  markerUnits = NULL,
  orient = NULL,
  preserveAspectRatio = NULL,
  viewBox = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mask"></a>}}
\subsection{Method \code{mask()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$mask(
  ...,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  maskUnits = NULL,
  maskContentUnits = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mpath"></a>}}
\subsection{Method \code{mpath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$mpath(..., xlink_href = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$path(..., d = NULL, pathLength = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pattern"></a>}}
\subsection{Method \code{pattern()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$pattern(
  ...,
  id = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  href = NULL,
  patternUnits = NULL,
  patternTransform = NULL,
  preserveAspectRatio = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radialGradient"></a>}}
\subsection{Method \code{radialGradient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$radialGradient(
  ...,
  id = NULL,
  cx = NULL,
  cy = NULL,
  r = NULL,
  fx = NULL,
  fy = NULL,
  fr = NULL,
  href = NULL,
  gradientUnits = NULL,
  gradientTransform = NULL,
  spreadMethod = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rect"></a>}}
\subsection{Method \code{rect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$rect(
  ...,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  rx = NULL,
  ry = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-script"></a>}}
\subsection{Method \code{script()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$script(..., type = NULL, href = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$set(..., to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$stop(..., offset = NULL, stop_color = NULL, stop_opacity = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-style"></a>}}
\subsection{Method \code{style()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$style(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-switch"></a>}}
\subsection{Method \code{switch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$switch(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-symbol"></a>}}
\subsection{Method \code{symbol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$symbol(
  ...,
  id = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  refX = NULL,
  refY = NULL,
  preserveAspectRatio = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text"></a>}}
\subsection{Method \code{text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$text(
  ...,
  x = NULL,
  y = NULL,
  dx = NULL,
  dy = NULL,
  rotate = NULL,
  textWidth = NULL,
  lengthAdjust = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textPath"></a>}}
\subsection{Method \code{textPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$textPath(
  ...,
  href = NULL,
  path = NULL,
  method = NULL,
  side = NULL,
  spacing = NULL,
  startOffset = NULL,
  textLength = NULL,
  lengthAdjust = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-title"></a>}}
\subsection{Method \code{title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$title(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-use"></a>}}
\subsection{Method \code{use()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$use(
  ...,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  href = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrix"></a>}}
\subsection{Method \code{matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$matrix(a, b, c, d, e, f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\subsection{Method \code{translate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$translate(x, y = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$scale(x, y = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rotate"></a>}}
\subsection{Method \code{rotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$rotate(a, x = NULL, y = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-skewX"></a>}}
\subsection{Method \code{skewX()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$skewX(a)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-skewY"></a>}}
\subsection{Method \code{skewY()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$skewY(a)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polygon"></a>}}
\subsection{Method \code{polygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$polygon(xs = NULL, ys = NULL, points = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polyline"></a>}}
\subsection{Method \code{polyline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$polyline(xs = NULL, ys = NULL, points = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize an SVGElement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$new(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{node name e.g. "circle"}

\item{\code{...}}{further arguments. Named arguments treated as attributes,
unnamed arguments treated as child nodes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\subsection{Method \code{update()}}{
Update the SVG Element.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$update(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{attributes and children to set on this node}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Named arguments are considered attributes and will overwrite
existing attributes with the same name. Set to NULL to delete the attribute

Unnamed arguments are appended to the list of child nodes.  These
should be text, other SVGElements or any ojbect that can be represented
as a single text string using "as.character()"

To print just the attribute name, but without a value, set to NA
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\subsection{Method \code{append()}}{
Append child nodes at the specified position.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$append(..., position = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{child nodes}

\item{\code{position}}{by default at the end of the list of children nodes
but 'position' argument can be used to set location by index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\subsection{Method \code{add()}}{
Simultaneous create an SVG element and add it as a child node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$add(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of node to create}

\item{\code{...}}{attributes and children of this newly created node}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
In contrast to most other methods, \code{$add()} returns
the newly created element, \emph{not} the document
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\subsection{Method \code{remove()}}{
Remove child objects at the given indicies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$remove(indicies)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indicies}}{indicies of the children to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_transform"></a>}}
\subsection{Method \code{reset_transform()}}{
Remove any transform attributes from this node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$reset_transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_character_inner"></a>}}
\subsection{Method \code{as_character_inner()}}{
Recursively convert this SVGElement and children to text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$as_character_inner(..., depth = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}

\item{\code{depth}}{recursion depth}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
single character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_character"></a>}}
\subsection{Method \code{as_character()}}{
Recursively convert this SVGElement and children to text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$as_character(..., depth = 0, include_declaration = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}

\item{\code{depth}}{recursion depth. default: 0}

\item{\code{include_declaration}}{Include the leading XML declaration? default: FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
single character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print the SVG string to the terminal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$print(include_declaration = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_declaration}}{Include the leading XML declaration? default: FALSE}

\item{\code{...}}{Extra arguments passed to \code{SVGElement$as_character()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
Save the text representation of this node and its children
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$save(filename, include_declaration = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename}

\item{\code{include_declaration}}{Include the leading XML declaration? default: FALSE}

\item{\code{...}}{Extra arguments passed to \code{SVGElement$as_character()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\subsection{Method \code{copy()}}{
Make a deep copy of this node and its children
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$copy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVGElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
